requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
htmltemplate
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
altair==4